Break up secure_dir_utils
