+++ coldfront/core/allocation/utils_/secure_dir_utils/README.md
@@ -1,0 +1,3 @@
+# README
+
+Utilities in secure_dir_utils.py should be broken up into smaller modules here.
